Add `TweenCompleted` event
Add a mechanism to `Tween<T>` to raise a Bevy event once the tween
completed. The event raising can be enabled with `set_completed_event()`
and `with_completed_event()`. The `TweenCompleted` event contains the
`Entity` on which the `Animator<T>` or `AssetAnimator<T>` owning the
tween are attached.

Update the `sequence` example to show how to use that new event
mechanism.