Upgrade to Bevy 0.8
Upgrade to the latest released version 0.8.0 of Bevy.

Add a new `bevy_asset` feature to enable animation of Bevy assets
(types implementing the `Asset` trait). Bevy 0.8 does not contain
`bevy_asset` in its defaul features, so this new feature reflects that
new possibility to build Bevy and í½ Bevy Tweening without support
for the `bevy_asset` crate. The new feature is enabled by default for
discoverability and to prevent a behavior breaking change.