Force `bevy_math/curve` feature through `bevy_color`
The `EaseFunction` definition depends on the `bevy_math/curve` feature,
which is unfortunately not re-exposed by Bevy. As a workaround, to make
sure that feature is always active (since `EaseFunction` is a core type
of `bevy_tweening`), we force a dependency on `bevy_color`, which itself
already forces that feature. This is not ideal and brittle, but good
enough until Bevy provides a better way.